+++ presidio-anonymizer/presidio_anonymizer/operators/hash.py
@@ -1,6 +1,6 @@
 """Hashes the PII text entity."""
 
-from hashlib import md5, sha256, sha512
+from hashlib import blake2b, sha256, sha512
 from typing import Dict
 
 from presidio_anonymizer.operators import Operator, OperatorType
@@ -8,12 +8,12 @@ from presidio_anonymizer.services.validators import validate_parameter_in_range
 
 
 class Hash(Operator):
-    """Hash given text with sha256/sha512/md5 algorithm."""
+    """Hash given text with sha256/sha512/blake2b algorithm."""
 
     HASH_TYPE = "hash_type"
     SHA256 = "sha256"
     SHA512 = "sha512"
-    MD5 = "md5"
+    BLAKE2B = "blake2b"
 
     def operate(self, text: str = None, params: Dict = None) -> str:
         """
@@ -25,14 +25,14 @@ class Hash(Operator):
         hash_switcher = {
             self.SHA256: lambda s: sha256(s),
             self.SHA512: lambda s: sha512(s),
-            self.MD5: lambda s: md5(s),
+            self.BLAKE2B: lambda s: blake2b(s, digest_size=20),
         }
         return hash_switcher.get(hash_type)(text.encode()).hexdigest()
 
     def validate(self, params: Dict = None) -> None:
         """Validate the hash type is string and in range of allowed hash types."""
         validate_parameter_in_range(
-            [self.SHA256, self.SHA512, self.MD5],
+            [self.SHA256, self.SHA512, self.BLAKE2B],
             self._get_hash_type_or_default(params),
             self.HASH_TYPE,
             str,